Check_EV_Properties.R

Uses COM objects to run Echoview and get transducer properties. These properties are saved to a .csv file in the Data_Compile_dir of the DirNameFile excel spreadsheet. The name defaults to TransducerConfigCheck(SurveyName).csv This works for EV files that have an ek5 sub-directory too. The ek5 subdirectory must be named 'ek5 EV'

Usage

Check_EV_Properties(SurveyName, DirNameFile, variables, sheet = case, Xnum = 2)

Arguments

SurveyName

name of survey as in excel file

DirNameFile

excel file name of paths

variables

- variables to export

sheet

is an optional argument to denote which sheet of the DirNameFile want to use. This defaults to the first sheet if not set.

Xnum

- the transducer number. Defaults to 2 (which is often 38 kHz)

database

is an optional argument for export format. Default is database (1). Spreadsheet =0

Examples

 if (FALSE) { # \dontrun{
Check_EV_Properties(SurveyName1in, DirNameFile,variables,sheet=case)
} # }