age1_indx_comp.R

Function to compare the age-1 abundance estimate from cases to that from the original age-1 index.

Usage

age1_indx_comp(
  SurveyNamein,
  DirNameFilein,
  imagedirin,
  casesin,
  hakeflagin,
  yearin,
  age1indfnamein,
  ...
)

Arguments

hakeflagin

adult hake (hakeflag=2), not filtering for hake (hakeflag=-1), age-0 hake (hakeflag=0), age-1 hake only (hakeflag=1), all hake (hakeflag=3), age-1+ (hakeflag=4). Passing this although assuming this will be hakeflag=1

yearin

is the year that is being investigated

SurveyName

- name of survey

DirNameFile

excel file - contains paths and locations of files. Uses Export_dir as the directory that the exported files are in.

imagedir

- the directory for the images files to be placed in

cases

- is the set of cases to be compared

age1indfname

is the directory the age-1 index file is found in (file needs to be named 'Age1_Full_Time_Series'

Examples

 if (FALSE) { # \dontrun{
T<-age1_indx_comp(SurveyName, DirNameFile,imagedir,cases,hakeflagin, yearin, age1indfname)
} # }