export_integration_EVdir

function to export integration from variables in a set of EV files. There are two excel files for inputs (paths, parameters to export). The exports go into a directory that is automatically named by date. This directory is in the export directory listed in the DirNameFile. The EVfile is not saved in the process. This works for EV files that have an ek5 sub-directory too. The ek5 subdirectory must be named 'ek5 EV'

Usage

export_integration_EVdir(
  Survey_Name,
  DirNameFile,
  variables,
  paramnamefile = NA,
  ni = 1,
  gdepth = 10,
  database = 1,
  sheet,
  datedir = 1,
  ...
)

Arguments

DirNameFile

excel file name of paths

variables

- variables to export

paramnamefile

optional excel file name of parameters to export or not export. The column ToExport is parameters to export. The column NoExport is parameters to not export

ni

is an optional argument for which file to start with (default ni=1 is the first file). This defaults to 1 (all EV files)

gdepth

is an optional argument for the cell depth. gdepth defaults to 10 m

database

is an optional argument for export format. Default is database (1). Spreadsheet =0

sheet

is an optional argument to denote which sheet of the DirNameFile want to use. This defaults to the first sheet if not set.

datedir

Flag to state whether to create and use a separate date sub-directory when redoing exports, so originals don't get overwritten. This defaults to using a date sub-directory (datedir=1 is default; make date directory. datedir=0 does not create a sub-directory)

SurveyName

name of survey as in excel file